reddit-bot/requirements.txt
praw>=7.7.0
google-cloud-aiplatform>=1.38.0
google-cloud-firestore>=2.14.0
python-dotenv>=1.0.0
